bridger/relayer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
web3==6.11.3
eth-account==0.10.0
